Fixed getuser by ID or name

+++ src/main/java/com/revature/pokemondb/services/UserService.java
@@ -4,7 +4,7 @@ import com.revature.pokemondb.exceptions.UsernameAlreadyExistsException;
 import com.revature.pokemondb.models.User;
 
 public interface UserService {
-	public User getUserById (int id);
+	public User getUserById (Long id);
 	
 	public User getUserByUsername (String username);
